move calsub updates to GORM
