+++ calsub/store.go
@@ -197,51 +197,6 @@ func (s *Store) CreateTx(ctx context.Context, tx *sql.Tx, cs *Subscription) (*Su
 	return n, err
 }
 
-// FindOneForUpdateTx will return a CalendarSubscription for the given userID that is locked for updating.
-func (s *Store) FindOneForUpdateTx(ctx context.Context, tx *sql.Tx, userID, id string) (*Subscription, error) {
-	err := permission.LimitCheckAny(ctx, permission.MatchUser(userID))
-	if err != nil {
-		return nil, err
-	}
-	err = validate.Many(
-		validate.UUID("ID", id),
-		validate.UUID("UserID", userID),
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	var cs Subscription
-	row := wrapTx(ctx, tx, s.findOneUpd).QueryRowContext(ctx, id, userID)
-	err = cs.scanFrom(row.Scan)
-	if err != nil {
-		return nil, err
-	}
-
-	return &cs, nil
-}
-
-// UpdateTx updates a calendar subscription with given information.
-func (s *Store) UpdateTx(ctx context.Context, tx *sql.Tx, cs *Subscription) error {
-	err := permission.LimitCheckAny(ctx, permission.MatchUser(cs.UserID))
-	if err != nil {
-		return err
-	}
-
-	n, err := cs.Normalize()
-	if err != nil {
-		return err
-	}
-
-	cfgData, err := json.Marshal(n.Config)
-	if err != nil {
-		return err
-	}
-
-	_, err = wrapTx(ctx, tx, s.update).ExecContext(ctx, cs.ID, cs.UserID, cs.Name, cs.Disabled, cfgData)
-	return err
-}
-
 // FindAllByUser returns all calendar subscriptions of a user.
 func (s *Store) FindAllByUser(ctx context.Context, userID string) ([]Subscription, error) {
 	err := permission.LimitCheckAny(ctx, permission.MatchUser(userID))
